fix: ignore .ssh pub keys (#109)
Signed-off-by: Justin Alvarez <alvajus@amazon.com>

Issue #, if available: Fixes #103. Still having trouble reproducing #38,
but possibly related

*Description of changes:*
- Our Lima configuration (finch.yaml) was using the default value for
the `ssh.loadDotSSHPubKeys` option, which is `true`. This may cause
errors if the user has any issues with the keys in their `.ssh`
directory. Since the correct key is generated and stored in Lima's
directories, and the user is not expected to interact directly with the
VM anyway, we can just set this to `false` explicitly
- In the future, we can consider adding this to finch's config (probably
by adding logic to `pkg/config`'s `lima_config_applier` and
`defaults.go`)

*Testing done:*
- Reproduced the error by creating an invalid pub key entry. Made the
change to my local `lima.yaml` and rebooted the vm, and did not see the
error


- [x] I've reviewed the guidance in CONTRIBUTING.md


#### License Acceptance

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.

Signed-off-by: Justin Alvarez <alvajus@amazon.com>