ci: automate step to update finch-core dependency in makefile (#136)
Signed-off-by: Vishwas Siravara <siravara@amazon.com>

Issue #, if available:

*Description of changes:*
Historically releasing `finch` requires us to manually update the
`finch-core` reference in the
[Makefile](https://github.com/runfinch/finch/blob/main/Makefile#L24).
This process is error prone and time consuming. This change automates
this step to a certain degree and does not require the user to modify
the Makefile manually. This change will eventually be triggered by a
release in finch-core, current trigger is `workflow_dispatch` and the
action `Update dependencies` has to be manually triggered before the
release of finch.
*Testing done:*
Manually executing the steps in the workflow on a remote server. 


- [X] I've reviewed the guidance in CONTRIBUTING.md


#### License Acceptance

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.

Signed-off-by: Vishwas Siravara <siravara@amazon.com>