feat: add workflow for uploading installers to release assets (#208)
Issue #, if available:

*Description of changes:*
This workflow is to be run after the `release-installer.yaml` workflow
completes. And after this workflow completes, the next step is to run
the `release-homebrew.yaml` workflow. All the three workflows should be
merged together once this
[issue](https://github.com/runfinch/finch/issues/106) is fixed.
*Testing done:*
Tested the workflow within my own fork.


- [x] I've reviewed the guidance in CONTRIBUTING.md


#### License Acceptance

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.

---------

Signed-off-by: Anqi Pang <anqipans@amazon.com>