fix: correctly handle local environment value pass-through (#158)
The `-e`,`--env` flag as well as values within an `--env-file` may
contain only a variable name with no "={VALUE}" portion. These entries
provide a shorthand to say "pass the existing environment value of this
variable into the container". Because of the VM boundary we need to
extrapolate these values on the Finch side and pass them as discrete
values into the Lima VM on the nerdctl command line. Also the file
referenced by `--env-file` may not be accessible inside the VM, so we
translate each entry in the file into `-e` entries on the command line
rather than fail on the VM side, unable to locate the file being
referenced.

Signed-off-by: Phil Estes <estesp@amazon.com>

Issue #, if available: Fixes #35

- [x] I've reviewed the guidance in CONTRIBUTING.md

#### License Acceptance

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.

---------

Signed-off-by: Phil Estes <estesp@gmail.com>