feat: Support special IP host-gateway in --add-host flag (#216)
Issue #, if available:
https://github.com/runfinch/finch/issues/209

*Description of changes:*
Support special IP `host-gateway` in `--add-host` flag. Finch will
resolve to the IP that can be used to access host from containers. It is
not configurable for now. See
[issue](https://github.com/runfinch/finch/issues/209) for more details.

*Testing done:*
unit tests and e2e tests
https://github.com/runfinch/common-tests/pull/29

- [ X ] I've reviewed the guidance in CONTRIBUTING.md


#### License Acceptance

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.

---------

Signed-off-by: Ziwen Ning <ningziwe@amazon.com>