fix: Correctly set and pick up environment variables (#315)
Issue #, if available:
Currently the environment variable is set in `/root/.bashrc` so that it
is [not easily
accessible](https://superuser.com/questions/1254422/linux-sudo-env-does-not-have-environment-variable-while-root-user-has)
by regular users even by specifying `sudo -E`. Normally we would access
the VM by regular user. In order to pick up the environment variables,
we need to store the vars in the /home/<user>.linux/.bashrc and added -E
to sudo command as sudo itself won't preserve the environment variables.

*Description of changes:*
Added -E to sudo command when we call limactl command.
Update the profile path so that env vars are written in
/home/<user>.linux/.bashrc.
Added e2e tests to make sure the content is written correctly into the
~/.finch/.config.json file in host system.
*Testing done:*
Manual testing and E2E tests


- [x] I've reviewed the guidance in CONTRIBUTING.md


#### License Acceptance

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.

---------

Signed-off-by: Ang Zhou <angzh@amazon.com>