feat: add experimental support for Cosign signing and verification (#296)
https://github.com/runfinch/finch/issues/295

*Description of changes:*
Add COSIGN_PASSWORD env pass-through to allow users use Cosign.

The feature should be experimental as it is experimental in Nerdctl. As
Finch points to Nerdctl documentation today, users could see Cosign
feature is experimental in Nerdctl documentation so will not mention
experimental in Finch explicitly.

*Testing done:*
The tests only covers signing by push and verification by pull and run
with the keys as MVP. The tests won't pass until the
[changed](https://github.com/containerd/nerdctl/pull/2109) is integrated
to Finch. Tested locally that it can work with this change in Nerdctl.


- [ X ] I've reviewed the guidance in CONTRIBUTING.md


#### License Acceptance

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.

Signed-off-by: Ziwen Ning <ningziwe@amazon.com>