feat: ECR credential integration into Finch  (#462)
Issue #, if available: 116

*Description of changes:*
Integrating ECR credential helper into Finch. 

**To Install the Credential Helper**
- The ECR credential helper can be setup with minimal configuration by
setting the `creds_helpers` option in `finch.yaml` with the value `-
ecr-login`.

- Once the option has been set the credential helper will be installed
on either `finch vm init ` or `finch vm start`. The binary will be
downloaded on the host machine and a `config.json` will be created and
populated in the `~/.finch/` folder if it doesn't already exist. If it
already exists, the value of `credsStore` will be overwritten to
`ecr-login` .

**How to stop using the credential helper with Finch**
- `config.json` needs to be either deleted, or the `credsStore`
parameter needs to have the value `ecr-login ` removed. Additionally the
creds_helper option in the finch.yaml needs to be removed.

- To fully remove the credential helper from the host machine, this can
be done by deleting it from the
`~/.finch/cred-helpers` folder.


The credential helper will support credentials from both the aws
credentials file and those that are stored as environment variables.


*Testing done:*
Added Unit Tests
Added Integration Test 


- [ x] I've reviewed the guidance in CONTRIBUTING.md


#### License Acceptance

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.

---------

Signed-off-by: kiryl1 <kirylkul@gmail.com>