ci: add script to update rootfs URL
for Finch on Windows #492, we need to provide a rootfs to WSL2. This
rootfs lives in our dependencies bucket, and will be updated from
time-to-time for security patches, bug fixes, etc. This commit will
automatically pull the most recent rootfs from the depenedencies bucket
as part of the Update Deps action.

Signed-off-by: Gavin Inglis <giinglis@amazon.com>