fix: Restart buildkit after containerd when provisioning (#461)
Issue #, if available: Fixes #412 

*Description of changes:*
- Previously, only containerd was restarted after configuring it to use
the data on the persistent disk. This changes the UUID of the server
worker. BuildKit also needs to be restarted to use the proper UUID. See
issue for why this is important.

*Testing done:*
- Local testing


- [x] I've reviewed the guidance in CONTRIBUTING.md


#### License Acceptance

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.

---------

Signed-off-by: Justin Alvarez <alvajus@amazon.com>