fix: option to use installed lima for SOCI e2e tests
to verify SOCI snapshotter is working correctly, limactl is used to
shell into the Finch VM and check that mounted snapshots match the
expected created by SOCI. In our nightly builds we test against an
installed version of Finch, but the previous SOCI e2e tests did not
account for this. Now, branch on `installed` to determine which lima
path to use.

Signed-off-by: Gavin Inglis <giinglis@amazon.com>