fix: use correct URL for rootfs (#589)
Issue #, if available:

Mismatch in rootfs between finch and finch-core on respective windev
branches - both in the version and in the file referenced. finch-core is
correctly providing full .tar.gz extension and latest rootfs.

Evident in jobs:

*
https://github.com/runfinch/finch/actions/runs/6386280972/job/17332665289?pr=584
*
https://github.com/runfinch/finch/actions/runs/6386379434/job/17332965393?pr=584

*Description of changes:*

*Testing done:*

Validated on windows dev instance. 

- [x] I've reviewed the guidance in CONTRIBUTING.md


#### License Acceptance

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.

Signed-off-by: Gavin Inglis <giinglis@amazon.com>