feat: windows persistent disk and other misc fixes/refactors (#594)
Issue #, if available:

*Description of changes:*
- Adds persistent disk support to Windows
- This took more work than anticipated, in order to deal with non-Admin
users
- `pkg/disk/dpgo/` is brand new code, and should be a focus of the
review
- `pkg/winutil/run_windows.go` is also new and requires careful review.
This is what allows Finch to run as the regular user, except for when it
needs Admin access to call `diskpart` (to create the persistent disk)
- Fix paths in `nerdctl_config_applier` to make the post-boot/init
shelling work
- Added winres to allow the finch.exe to have metadata attached to it.
This is WIP, need final icons and descriptions etc.
- Large (in terms of lines changed) refactor of `pkg/path/finch.go`, but
it should have no impact on functionality (needs careful review)
- Fixed the Makefile's `clean` target for Windows
- Most of the other changes are just noise from refactoring (like,
literally renaming things). Nothing major, but take a look if possible.
Sorry the diff is so large

*Testing done:*

- [x] I've reviewed the guidance in CONTRIBUTING.md

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.

---------

Signed-off-by: Vishwas Siravara <siravara@amazon.com>
Signed-off-by: Justin Alvarez <alvajus@amazon.com>
Signed-off-by: Vishwas Siravara <vsiravar@gmail.com>
Co-authored-by: Vishwas Siravara <siravara@amazon.com>
Co-authored-by: Vishwas Siravara <vsiravar@gmail.com>
Signed-off-by: Vishwas Siravara <siravara@amazon.com>