ci: pin actions (#605)
Issue #, if available:

*Description of changes:*
  - Pinned all actions to commit SHAs
    - Just used the latest available version
- Dependabot should actually update the version comment too, thanks to
[this
feature](https://github.com/dependabot/dependabot-core/issues/4691),
which is pretty neat
- `Homebrew/actions/setup-homebrew` isn't really "vended" and the
recommendation is to just use the latest master commit of the repo,
which is why the version name just says "master"
- `WyriHaximus/github-action-get-previous-tag` hasn't been updated in a
while and the functionality doesn't seem that hard to replicate on our
own. Might be good to remove it in the future

*Testing done:*

- [x] I've reviewed the guidance in CONTRIBUTING.md

#### License Acceptance

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.

Signed-off-by: Justin Alvarez <alvajus@amazon.com>
Signed-off-by: Vishwas Siravara <siravara@amazon.com>