fix: unmarshal additional disks
disks in the list format in lima override.yaml were not being
unmarshalled properly, causing a disk with the name "" to be mixed with
finch lima yaml, resulting in startup failure. This only occurs on inits
of a vm other than the very first init. init a vm the first time,
_output/lima/data/_config/override.yaml:

```yaml
additionalDisks:
    - name: finch
```

stop and remove the vm, then re-init:
```yaml
additionalDisks:
    - name: ""
    - name: finch
```

to fix, copy upstream logic from lima to use a custom unmarshaler for
the Disk type.

Signed-off-by: Gavin Inglis <giinglis@amazon.com>