ci: Make unsigned MSI has timestamp in its name to easily unique fetch it back (#636)
*Description of changes:*
When uploading the unsigned to S3, the signer will assign a random
postfix to the original MSI.
However, since 0.0.0 is a common test install name, it is hard to
uniquely identify the signed MSI since there can be historical ones.
Added timestamp in the name so that it can be easily fetched back.

*Testing done:*
Tested with this action run:
https://github.com/runfinch/finch/actions/runs/6503614023/job/17664348385

- [X] I've reviewed the guidance in CONTRIBUTING.md

#### License Acceptance

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.

---------

Signed-off-by: cnkevin@amazon.com <chaoningusc@gmail.com>
Co-authored-by: cnkevin@amazon.com <chaoningusc@gmail.com>
Signed-off-by: Vishwas Siravara <siravara@amazon.com>