fix: Fix to remove the virtual machine storage in deps/finch-core/downloads/os by make clean
Issue #, if available: #537

*Description of changes:*

In finch, when booting a virtual machine using Lima, it is necessary to
download the virtual machine storage in QCOW2 format.

The QCOW2 virtual machine storage thus downloaded and located in
deps/finch-core/downloads/os is not deleted when make clean is run in the
home directory as currently implemented.

This file may be larger than 500 MB in size and is reported in issue/#537
such that it is expected to be deleted when make clean is run.

Therefore, this fix will remove the QCOW2 format downloaded files in
deps/finch-core/downloads/os when make clean is run in the home directory.

*Testing done:* yes

- [x] I've reviewed the guidance in CONTRIBUTING.md

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.

---------

Signed-off-by: Hayato Kiwata <haytok@amazon.co.jp>