fix: Fix to be able to run finch build with --ssh option
In the current implementation, the following error occurs when running
finch build with the `--ssh default` option. This issue is also reported
in issue/452.

  ```
  error: invalid empty ssh agent socket: make sure SSH_AUTH_SOCK is set
  FATA[0000] no image was built
  FATA[0000] exit status 1
  ```

This is because the ssh agent is not starting on the vm started using lima
and the SSH_AUTH_SOCK variable is empty. As a result, this error occurs.

On the other hand, setting forwardAgent to true in finch.yaml will
configure the vm to forward the ssh agent.

As a result, the ssh agent will be started in the vm and finch build can
be executed using the `--ssh default` option.

Therefore, this commit will fix it so that finch build can be run using
the `--ssh default` option.

Signed-off-by: Hayato Kiwata <haytok@amazon.co.jp>