fix: Fix to delete ~/.finch when uninstalling finch
In the current implementation, uninstalling finch does not delete the
~/.finch directory.

This directory may hold several GB of container data disks, which can
consume a lot of space and be left unremoved.

This issue has also been reported in issue/#457 and needs to be fixed.

Therefore, this fix will ensure that when finch is uninstalled, ~/.finch
is also deleted to free up disk space for finch.

Signed-off-by: Hayato Kiwata <haytok@amazon.co.jp>