fix: Fix to allow users to configure disk size for finch space
In the current implementation, the disk space allocated for finch is
hard-coded with a fixed value of 50G. Therefore, when users pull large
docker images and use more than 50G of the allocated finch space,
"no space left on device" error may occur.

As mentioned above, the disk size available for finch is fixed at 50GB and
cannot be set by users. To work around this setting, they need to rewrite
the hard-coded parts and use their own build of finch, but this is very
time-consuming.

Alternatively, they can use the following method to change the disk size.

- https://github.com/runfinch/finch/issues/275#issuecomment-1577010453

Thus, although there is some workarounds to adjust the disk size for
finch, users will be able to use finch more flexibly if they are able to
specify the disk size that finch can use at the time of vm init.

Therefore, this fix improves usability by allowing users to run "finch vm
init" and flexibly set the size of the disk created for finch based on the
finchDiskSize in ~/.finch/finch.yaml.

Note that for finch at present, the size of finchDiskSize becomes the disk
size of the finch area only when the vm is initialized ("finch vm init").

Signed-off-by: Hayato Kiwata <haytok@amazon.co.jp>