ci: update release-please action config to work with v4 (#738)
Issue #, if available: N/A

*Description of changes:* Our release-please GH action version was
recently bumped from v3 to v4 by a dependabot PR merge. Since then,
we've merged subsequent PRs but haven't had a new automated release PR
be created by the bot. This seems to be caused by the [breaking changes
described
here](https://github.com/google-github-actions/release-please-action?tab=readme-ov-file#upgrading-from-v3-to-v4).

This PR tries to migrate our config to work with version 4 of
release-please.

*Testing done:*
- It's kinda hard to test this since the action runs remotely and uses
the GitHub repo itself as input. If anyone has an idea of how to test
this that can be done _before_ merging this PR, lmk and I can try it.
Otherwise, seems pretty low-risk to merge and see what happens.


- [x] I've reviewed the guidance in CONTRIBUTING.md


#### License Acceptance

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.

Signed-off-by: Justin Alvarez <alvajus@amazon.com>