fix(windows-installer): uninstall script behavior (#822)
Issue #, if available:
- After my recent changes to the uninstall script (which were aimed at
making sure Lima was not running and the WSL2 instance was stopped), I
tested the new installer, and it still had weird behavior. I noticed
that the script would exit immediately if the Lima stop command failed,
so I added the `&` separators. After that, I noticed that the WSL
commands would fail immediately. This was because the commands were
being executed as a Local install user, instead of the target user.
Changing `Impersonate` to `yes` should resolve that issue
([docs](https://wixtoolset.org/docs/v3/xsd/wix/customaction/)).

_Description of changes:_
- Ensure that all commands execute by separating them with & instead of
letting failures stop the script
- Ensure script runs in the user's context by setting `Impersonate` to
`yes`

*Testing done:*
- Local testing


- [x] I've reviewed the guidance in CONTRIBUTING.md


#### License Acceptance

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.

Signed-off-by: Justin Alvarez <alvajus@amazon.com>