fix: add finch attach command
The finch attach command cannot be executed at the top level.
Therefore, when the finch attach -h command is executed, the following
error occurs.

  ```
  > finch attach -h
  FATA[0000] unknown command "attach" for "finch"
  ```

Also, this bug has been reported in issue/747 below, and it is hoped that
finch attach will be executed.

- https://github.com/runfinch/finch/issues/747

Meanwhile, with the nerdctl command, it is possible to execute the nerdctl
attach command.

Therefore, in this commit, we will make corrections so that the finch
attach command can be executed.

Signed-off-by: Hayato Kiwata <haytok@amazon.co.jp>