docs: Add docs for finch vm settings command
In my previous pull request, we added the functionality to change the
number of CPUs and memory size allocated to VMs.

  - https://github.com/runfinch/finch/pull/887

However, at that time, we did not add documentations for the finch vm
settings command.

Therefore, in this fix, we will add documentations for the finch vm
settings command.

Signed-off-by: Hayato Kiwata <haytok@amazon.co.jp>