ci: Add a CI to check documentation changes
This fix introduces a new CI to check whether the command reference
documentations need to be updated.

When new commands or options are added, the
`finch gen-docs generate -p cmd/finch/` command should be run locally to
update the documentation.

However, it's easy to forget this step.

An issue has also been created that proposes to detect these differences.

- https://github.com/runfinch/finch/issues/942

Therefore, in this fix, the new CI step will generate the documentation
and compare it with the existing files.
If there are differences, the CI will fail, prompting contributors to
update the documentations.

Signed-off-by: Hayato Kiwata <haytok@amazon.co.jp>