ci(refactor): split build-and-test-pkg for better reuse (#953)
Issue #, if available:

*Description of changes:*
While I was fixing the macOS pkg build error on macOS 11 (due to `yq`
from brew not building properly), I noticed that most of the file was
repeated (due to different archs being built/tested as different jobs)
which caused it to be pretty unnecessarily unwieldy. This was the case
for both the build and the test workflow jobs.

I initially thought I could combine them using matrices, but then we
lose the ability to run the test jobs as soon as each respective build
completes (with a matrix, it would have to wait until _both_ builds
complete, which isn't even guaranteed). There were other complications
that made matrices hard, like the fact that we only want to build once
per arch, but we want to test on several other arch/version
combinations.

Instead, I was able to split the workflows into reusable/composite
workflows, so that only the inputs vary, and the script is mostly the
same.

*Testing done:*


- [x] I've reviewed the guidance in CONTRIBUTING.md


#### License Acceptance

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.

---------

Signed-off-by: Justin Alvarez <alvajus@amazon.com>