fix: Remove unnecessary error handling
Inside the ConfigureDefaultLimaYaml() function, there is a process to
write to the config file as follows:

    ```
    if err := afero.WriteFile(lca.fs, lca.limaDefaultConfigPath, limaCfgBytes, 0o600); err != nil {
    	return fmt.Errorf("failed to write to the lima config file: %w", err)
    }

    if err != nil {
    	return fmt.Errorf("unable to apply override config: %w", err)
    }
    ```

The current implementation has two error checks to determine if the file
can be written.

The err object in the second part is guaranteed to be nil due to the
previous if statement.

Therefore, this fix removes the unnecessary second error check.

Signed-off-by: Hayato Kiwata <haytok@amazon.co.jp>