ci(release): fix release build generation (#1007)
Issue #, if available:

*Description of changes:*
Recently in https://github.com/runfinch/finch/pull/969, the Makefile was
refactored significantly, and the mechanism that was used to override
the installation directory was removed ([used
here](https://github.com/runfinch/finch/blob/main/.github/workflows/build-pkg.yaml#L57)).
This just adds that mechanism back. There could be a better way to do
this, but for now just reverting back to how it was done before the
recent change.

*Testing done:*
Tested locally. The `_output/os/finch.yaml` file correctly used
`/Applications/Finch/`, which was the path I overwrote, as a prefix


- [x] I've reviewed the guidance in CONTRIBUTING.md


#### License Acceptance

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.

Signed-off-by: Justin Alvarez <alvajus@amazon.com>